requirements/requirements.txt
django
django-suit
djangorestframework
markdown
django-filter
pytz
